requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1699370045702/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
biopython==1.79
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1677499911949/work
brotlipy==0.7.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama @ file:///croot/colorama_1672386526460/work
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1691048088238/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1691009212940/work
configparser==5.3.0
cryptography @ file:///croot/cryptography_1677533068310/work
decorator==5.1.0
dgl==0.9.1.post1
DLLogger @ git+https://github.com/NVIDIA/dllogger@0540a43971f4a8a16693a9de9de73c1072020769
docker-pycreds==0.4.0
e3nn==0.3.3
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
flit_core @ file:///croot/flit-core_1679397103445/work/source/flit_core
future @ file:///croot/future_1677599870788/work
gitdb==4.0.10
GitPython==3.1.31
hydra-core==1.3.2
icecream @ file:///home/conda/feedstock_root/build_artifacts/icecream_1658422146206/work
idna @ file:///croot/idna_1666125576474/work
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1695536281965/work
jsonpointer @ file:///home/conda/feedstock_root/build_artifacts/jsonpointer_1695397261020/work
keopscore==2.2.3
libmambapy @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1699607298048/work/libmambapy
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mpmath==1.3.0
networkx @ file:///croot/networkx_1678964333703/work
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
omegaconf==2.3.0
opt-einsum==3.3.0
opt-einsum-fx==0.1.4
packaging @ file:///croot/packaging_1678965309396/work
pathtools==0.1.2
Pillow==9.4.0
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1693086607691/work
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
ProDy==2.4.1
promise==2.3
protobuf==4.22.1
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
pybind11==2.13.1
pycosat @ file:///home/conda/feedstock_root/build_artifacts/pycosat_1696355771368/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pykeops==2.2.3
pynvml==11.0.0
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.2
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
PyYAML==6.0
requests @ file:///croot/requests_1678709721434/work
-e git+https://github.com/geraseva/RFdiffusion@1d16cba8b82d38c25a16e75eeb3de10b4f97ef46#egg=rfdiffusion
ruamel.yaml @ file:///home/conda/feedstock_root/build_artifacts/ruamel.yaml_1699007328119/work
ruamel.yaml.clib @ file:///home/conda/feedstock_root/build_artifacts/ruamel.yaml.clib_1695996823774/work
scipy==1.10.1
se3-transformer==1.0.0
sentry-sdk==1.19.1
shortuuid==1.0.11
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
subprocess32==3.5.4
sympy==1.11.1
torch @ file:///home/conda/feedstock_root/build_artifacts/pytorch-recipe_1633418020348/work
torchaudio==0.9.0a0+a85b239
torchvision @ file:///croot/torchvision_1670313549871/work
tqdm @ file:///croot/tqdm_1679561862951/work
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3 @ file:///croot/urllib3_1680254681959/work
wandb==0.12.0
zstandard==0.22.0
